changing the LNS template interface to use unsigned int instead of size_t for non-type parameters

+++ include/universal/number/lns/attributes.hpp
@@ -16,14 +16,14 @@ namespace sw { namespace universal {
 // They should not be used for the core algorithms.
 
 	// free function sign
-template<size_t nbits, size_t rbits, typename bt, auto ...xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto ...xtra>
 bool sign(const lns<nbits, rbits, bt, xtra...>& v) {
 	return v.sign();
 }
 
 // generate the maxneg through maxpos value range of a logarithmic number system configuration
 // the type of arithmetic, Modulo or Saturating, does not affect the range
-template<size_t nbits, size_t rbits, typename bt, auto ...xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto ...xtra>
 std::string lns_range(const lns<nbits, rbits, bt, xtra...>& v) {
 	using LNS = lns<nbits, rbits, bt, xtra...>;
 	std::stringstream s;
