+++ include/universal/number/lns/math/trigonometry.hpp
@@ -12,61 +12,61 @@ namespace sw { namespace universal {
 // One radian is equivalent to 180/PI degrees
 
 // sine of an angle of x radians
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> sin(lns<nbits, rbits, bt, xtra...> x) {
 	return lns<nbits, rbits, bt, xtra...>(std::sin(double(x)));
 }
 
 // cosine of an angle of x radians
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> cos(lns<nbits, rbits, bt, xtra...> x) {
 	return lns<nbits, rbits, bt, xtra...>(std::cos(double(x)));
 }
 
 // tangent of an angle of x radians
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> tan(lns<nbits, rbits, bt, xtra...> x) {
 	return lns<nbits, rbits, bt, xtra...>(std::tan(double(x)));
 }
 
 // cotangent of an angle of x radians
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> atan(lns<nbits, rbits, bt, xtra...> x) {
 	return lns<nbits, rbits, bt, xtra...>(std::atan(double(x)));
 }
 		
 // Arc tangent with two parameters
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> atan2(lns<nbits, rbits, bt, xtra...> y, lns<nbits, rbits, bt, xtra...> x) {
 	return lns<nbits, rbits, bt, xtra...>(std::atan2(double(y),double(x)));
 }
 
 // cosecant of an angle of x radians
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> acos(lns<nbits, rbits, bt, xtra...> x) {
 	return lns<nbits, rbits, bt, xtra...>(std::acos(double(x)));
 }
 
 // secant of an angle of x radians
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> asin(lns<nbits, rbits, bt, xtra...> x) {
 	return lns<nbits, rbits, bt, xtra...>(std::asin(double(x)));
 }
 
 // cotangent an angle of x radians
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> cot(lns<nbits, rbits, bt, xtra...> x) {
 	return lns<nbits, rbits, bt, xtra...>(std::tan(sw::universal::m_pi_2-double(x)));
 }
 
 // secant of an angle of x radians
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> sec(lns<nbits, rbits, bt, xtra...> x) {
 	return lns<nbits, rbits, bt, xtra...>(1.0/std::cos(double(x)));
 }
 
 // cosecant of an angle of x radians
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> csc(lns<nbits, rbits, bt, xtra...> x) {
 	return lns<nbits, rbits, bt, xtra...>(1.0/std::sin(double(x)));
 }
